[refactor] conditionReducer 파일 분리
